feat: clean admin panel
