+++ src/main/java/fr/chatavion/server/ServerApplication.java
@@ -1,11 +1,9 @@
 package fr.chatavion.server;
 
+import fr.chatavion.server.admin.Console;
 import fr.chatavion.server.dns.MockDNS;
-import fr.chatavion.server.dns.util.Community;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.util.Scanner;
 
 @SpringBootApplication
 public class ServerApplication {
@@ -15,17 +13,6 @@ public class ServerApplication {
 
         new MockDNS(53).start();
 
-        console();
+        new Console().start();
 	}
-
-    public static void console() {
-        Scanner scanner = new Scanner(System.in);
-        while (scanner.hasNext()) {
-            var input = scanner.next();
-            if(input.startsWith("community")) {
-                var communityName = scanner.next();
-                Community.createCommunity(communityName);
-            }
-        }
-    }
 }
